don't delete certificate templates or badge icons if they are used
--HG--
branch : 3.6

+++ modules/admin/certbadge.php
@@ -44,19 +44,33 @@ $tool_content .= action_bar(array(
               'level' => 'primary-label')));
 
 if (isset($_GET['del_badge'])) { // delete badge icon
-    $badge_icon = Database::get()->querySingle("SELECT filename FROM badge_icon WHERE id = ?d", $_GET['del_badge'])->filename;
-    if (unlink($webDir . BADGE_TEMPLATE_PATH . $badge_icon)) {
-        Database::get()->query("DELETE FROM badge_icon WHERE id = ?d", $_GET['del_badge']);
-        Session::Messages($langDelWithSuccess, 'alert-success');
-    }    
+    $sql_badge_icon = Database::get()->querySingle("SELECT id, filename FROM badge_icon WHERE id = ?d", $_GET['del_badge']);
+    $badge_icon_id = $sql_badge_icon->id;
+    $cnt = Database::get()->querySingle("SELECT COUNT(*) AS cnt FROM badge WHERE icon = ?d", $badge_icon_id)->cnt;
+    if ($cnt > 0) {  // don't delete if it's used by a badge (foreing key constrain)
+        Session::Messages($langIconBelongsToBadge, 'alert-warning');
+    } else {
+        $badge_icon = $sql_badge_icon->filename;
+        if (unlink($webDir . BADGE_TEMPLATE_PATH . $badge_icon)) {
+            Database::get()->query("DELETE FROM badge_icon WHERE id = ?d", $_GET['del_badge']);
+            Session::Messages($langDelWithSuccess, 'alert-success');
+        }
+    }
 }
 
 if (isset($_GET['del_cert'])) { // delete certificate template
-    $cert_template = Database::get()->querySingle("SELECT filename FROM certificate_template WHERE id = ?d", $_GET['del_cert'])->filename;
-    if (unlink($webDir . CERT_TEMPLATE_PATH . $cert_template)) {
-        Database::get()->query("DELETE FROM certificate_template WHERE id = ?d", $_GET['del_cert']);
-        Session::Messages($langDelWithSuccess, 'alert-success');
-    }    
+    $sql_cert_template = Database::get()->querySingle("SELECT id, filename FROM certificate_template WHERE id = ?d", $_GET['del_cert']);
+    $cert_template_id = $sql_cert_template->id;
+    $cnt = Database::get()->querySingle("SELECT COUNT(*) AS cnt FROM certificate WHERE template = ?d", $cert_template_id)->cnt;
+    if ($cnt > 0) { // don't delete it if it's used by a certificate (foreign key constrain)
+        Session::Messages($langTemplateBelongsToCert, 'alert-warning');
+    } else {
+        $cert_template = $sql_cert_template->filename;
+        if (unlink($webDir . CERT_TEMPLATE_PATH . $cert_template)) {
+            Database::get()->query("DELETE FROM certificate_template WHERE id = ?d", $_GET['del_cert']);
+            Session::Messages($langDelWithSuccess, 'alert-success');
+        }
+    }
 }
 
 if (isset($_POST['submit_cert_template'])) { // insert certificate template
